requirements.txt
bleach==5.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
commonmark==0.9.1
cryptography==38.0.4
docutils==0.19
flake8==6.0.0
idna==3.4
importlib-metadata==5.1.0
jaraco.classes==3.2.3
jeepney==0.8.0
keyring==23.11.0
mccabe==0.7.0
more-itertools==9.0.0
mypy==0.991
mypy-extensions==0.4.3
pkginfo==1.8.3
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.16.0
pyflakes==3.0.1
Pygments==2.13.0
PyYAML==6.0
readme-renderer==37.3
requests==2.28.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==12.6.0
SecretStorage==3.3.3
six==1.16.0
tomli==2.0.1
twine==4.0.1
types-PyYAML==6.0.12.2
typing_extensions==4.4.0
urllib3==1.26.13
webencodings==0.5.1
zipp==3.11.0
